#236 fix : 금액 단위 변경(화폐가 달라지면 소수점이 필요함)

+++ src/main/java/com/floney/floney/analyze/dto/process/BookAnalyzer.java
@@ -7,26 +7,26 @@ import java.util.Map;
 
 public class BookAnalyzer {
     private static final int STANDARD = 0;
-    private final Map<String, Long> totalExpenses;
+    private final Map<String, Float> totalExpenses;
 
-    public BookAnalyzer(Map<String, Long> totalExpenses) {
+    public BookAnalyzer(Map<String, Float> totalExpenses) {
         this.totalExpenses = totalExpenses;
     }
 
-    public long differenceInAndOutCome() {
+    public float differenceInAndOutCome() {
         return totalExpenses.get(AssetType.INCOME.getKind()) - totalExpenses.get(AssetType.OUTCOME.getKind());
     }
 
-    public long calculateCurrentAsset(long initAsset, long difference) {
+    public float calculateCurrentAsset(float initAsset, float difference) {
         if (difference < STANDARD) {
             return initAsset - Math.abs(difference);
         }
         return initAsset + difference;
     }
 
-    public AnalyzeResponseByAsset analyzeAsset(long savedAsset) {
-        long difference = differenceInAndOutCome();
-        long currentAsset = calculateCurrentAsset(savedAsset, difference);
+    public AnalyzeResponseByAsset analyzeAsset(float savedAsset) {
+        float difference = differenceInAndOutCome();
+        float currentAsset = calculateCurrentAsset(savedAsset, difference);
 
         return AnalyzeResponseByAsset.of(difference, savedAsset, currentAsset);
     }
