+++ src/main/java/com/floney/floney/book/dto/process/DayLines.java
@@ -13,8 +13,8 @@ import java.util.stream.Collectors;
 @Getter
 @EqualsAndHashCode
 public class DayLines {
-    private final Long id;
-    private final Long money;
+    private final long id;
+    private final float money;
     private final String img;
     private final List<String> category;
     private final AssetType assetType;
@@ -24,7 +24,7 @@ public class DayLines {
     private final String userNickName;
 
     @Builder
-    public DayLines(Long id, Long money, String img, String content, List<String> category, AssetType assetType, String userEmail, boolean exceptStatus, String userNickName) {
+    public DayLines(Long id, float money, String img, String content, List<String> category, AssetType assetType, String userEmail, boolean exceptStatus, String userNickName) {
         this.id = id;
         this.money = money;
         this.img = img;
